audio: Output latency update in Audio Track
Audio Track needs to query latency from the output descriptor each
time it uses mAfLatency to calculate timestamp or return track
latency. Until now, the output descriptor latency is being queried
and stored when a track is created. However there are usecases like
when there is a device switch but not a track tear down, the path
latency changes but the track does not requery the output latency.
This fix updates the output latency in Audio Track such that the
latency being used reflects the correct device latency.

CRs-Fixed: 2044756
Change-Id: I5e59e03149b172dd148781f33d9f5527f389da01
Signed-off-by: Marco Zanin (B--B) <mrczn.bb@gmail.com>
Signed-off-by: Tab Fitts <tfitts1212@gmail.com>