media: Change the location of dumping stats
The latest photos apk calls the setVideoSurfaceTexture
with a null bufferProducer. Hence the surface is set as
NULL.
Due to this, the video decoder is shutdown before the
surface is set. Hence the stats are lost. Thus when the
app calls reset after surface is set, the stats are not
printed.

Change-Id: Iafc09c50840960be4a20e55b34d5d31ba6a2da49